LayoutTests/imported/w3c/web-platform-tests/IndexedDB/abort-in-initial-upgradeneeded.any-expected.txt


Harness Error (FAIL), message = Unhandled rejection: assert_equals: expected 0 but got 1

TIMEOUT An abort() in the initial onupgradeneeded sets version back to 0 Test timed out

Harness Error (FAIL), message = Unhandled rejection: assert_equals: expected 0 but got 1

TIMEOUT An abort() in the initial onupgradeneeded sets version back to 0 Test timed out
